# docs/openapi.json
{
  "components" : {
    "schemas" : {
...
      "DreamResponse" : {
        "properties" : {
          "completedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "error" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "observed" : {
            "type" : "string"
          },
          "observer" : {
            "type" : "string"
          },
          "produced" : {
            "format" : "int32",
            "type" : "integer"
          },
          "status" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ErrorResponse" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
      "EventResponse" : {
        "properties" : {
          "actor" : {
            "type" : "string"
          },
          "afterContent" : {
            "type" : "string"
          },
          "beforeContent" : {
            "type" : "string"
          },
          "conclusionId" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "detail" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "event" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "PeerCardResponse" : {
        "properties" : {
          "lines" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "observed" : {
            "type" : "string"
          },
          "observer" : {
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ScheduleDream" : {
        "properties" : {
          "observed" : {
            "minLength" : 1,
            "type" : "string"
          },
          "observer" : {
            "minLength" : 1,
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          }
        },
        "required" : [ "observed", "observer" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearer" : {
        "bearerFormat" : "JWT",
        "description" : "A token from `POST /v1/tokens`, in four nesting scopes: admin, workspace, peer, session. Each operation below names the narrowest one that may call it.",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Pair-scoped memory. Every fact belongs to a directed (observer, observed) pair: what bob remembers about alice is not what alice remembers about herself, and no route crosses that line.\n\nThree tiers of read: `context` assembles a session's summary and recent messages within a token budget, `recall` ranks conclusions by six fused signals, and `chat` runs an agentic loop over both when the answer needs more than one lookup.",
    "license" : {
      "name" : "Apache-2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "title" : "aimon-memory",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/v1/workspaces/{workspace}/conclusions/{id}/events" : {
      "get" : {
        "description" : "**Requires a `peer` token or wider.**",
        "operationId" : "history",
        "parameters" : [ {
          "in" : "path",
          "name" : "workspace",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 100,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/EventResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Malformed: a body that is not JSON, a field that fails validation, or a parameter of the wrong type."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "No bearer token, or one that does not verify."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The token verifies but is narrower than this route requires, or does not cover the workspace, peer or session named in the path."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "No such endpoint, or no such workspace, peer, session or conclusion. Also returned in place of 403 where telling the two apart would say whether another pair's row exists."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The request could not be completed. The body says no more than that, deliberately."
          }
        },
        "summary" : "Read a conclusion's audit trail",
        "tags" : [ "conclusions" ]
      }
    },
...
    "/v1/workspaces/{workspace}/dreams" : {
      "get" : {
        "description" : "**Requires a `workspace` token or wider.**",
        "operationId" : "list_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "workspace",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "observer",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "observed",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/DreamResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Malformed: a body that is not JSON, a field that fails validation, or a parameter of the wrong type."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "No bearer token, or one that does not verify."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The token verifies but is narrower than this route requires, or does not cover the workspace, peer or session named in the path."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "No such endpoint, or no such workspace, peer, session or conclusion. Also returned in place of 403 where telling the two apart would say whether another pair's row exists."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The request could not be completed. The body says no more than that, deliberately."
          }
        },
        "summary" : "List a pair's dreams",
        "tags" : [ "dreams" ]
      },
      "post" : {
        "description" : "**Requires a `workspace` token or wider.**",
        "operationId" : "schedule",
        "parameters" : [ {
          "in" : "path",
          "name" : "workspace",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ScheduleDream"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DreamResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Malformed: a body that is not JSON, a field that fails validation, or a parameter of the wrong type."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "No bearer token, or one that does not verify."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The token verifies but is narrower than this route requires, or does not cover the workspace, peer or session named in the path."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "No such endpoint, or no such workspace, peer, session or conclusion. Also returned in place of 403 where telling the two apart would say whether another pair's row exists."
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "A dream is already running for this pair. A partial unique index decides that, so a caller racing the automatic scheduler finds out here rather than from a stale read."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The request could not be completed. The body says no more than that, deliberately."
          }
        },
        "summary" : "Schedule a dream now",
        "tags" : [ "dreams" ]
      }
    },
    "/v1/workspaces/{workspace}/peer-card" : {
      "get" : {
        "description" : "**Requires a `peer` token or wider.**",
        "operationId" : "card",
        "parameters" : [ {
          "in" : "path",
          "name" : "workspace",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "observer",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "observed",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PeerCardResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Malformed: a body that is not JSON, a field that fails validation, or a parameter of the wrong type."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "No bearer token, or one that does not verify."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The token verifies but is narrower than this route requires, or does not cover the workspace, peer or session named in the path."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "No such endpoint, or no such workspace, peer, session or conclusion. Also returned in place of 403 where telling the two apart would say whether another pair's row exists."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The request could not be completed. The body says no more than that, deliberately."
          }
        },
        "summary" : "Get a pair's peer card",
        "tags" : [ "dreams" ]
      }
    },
    "/v1/workspaces/{workspace}/peer-card/refresh" : {
      "post" : {
        "description" : "**Requires a `peer` token or wider.**",
        "operationId" : "refresh",
        "parameters" : [ {
          "in" : "path",
          "name" : "workspace",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "observer",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "observed",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PeerCardResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Malformed: a body that is not JSON, a field that fails validation, or a parameter of the wrong type."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "No bearer token, or one that does not verify."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The token verifies but is narrower than this route requires, or does not cover the workspace, peer or session named in the path."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "No such endpoint, or no such workspace, peer, session or conclusion. Also returned in place of 403 where telling the two apart would say whether another pair's row exists."
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The request conflicts with existing data, or names a workspace, peer or session that does not exist. The database refused it, and the reply says no more than that: the driver's message quotes the constraint, the table and often the value."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "The request could not be completed. The body says no more than that, deliberately."
          },
          "503" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "No model provider is configured."
          }
        },
        "summary" : "Regenerate the peer card now",
        "tags" : [ "dreams" ]
      }
    },
...
  },
  "security" : [ {
    "bearer" : [ ]
  } ],
  "servers" : [ {
    "description" : "Relative to wherever this service is deployed.",
    "url" : "/"
  } ],
...
}